% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dist.R
\name{calc_dist}
\alias{calc_dist}
\title{Calculate string distance matrix}
\usage{
calc_dist(x, ...)
}
\arguments{
\item{x}{Character vector of medication names to be standardised}

\item{...}{Additional arguments to \code{stringdistmatrix}}
}
\value{
Matrix of string distances
}
\description{
Calculate string distance matrix
}
\examples{
drugs <- c("Paracetamol", "Paracetomol")
calc_dist(drugs)
}
