% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pain.med-package.R
\docType{package}
\name{pain.med-package}
\alias{pain.med}
\alias{pain.med-package}
\title{pain.med: "Handle Pain Medication Data"}
\description{
"Supports handling of pain medication names."
}
\author{
\strong{Maintainer}: Paul Newell \email{p.newell@exeter.ac.uk} (\href{https://orcid.org/0009-0001-6071-6209}{ORCID})

}
\keyword{internal}
