% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise.R
\name{standardise}
\alias{standardise}
\title{Standardise pain medication names}
\usage{
standardise(x, ..., quiet = FALSE)
}
\arguments{
\item{x}{Character vector of names to be standardised}

\item{...}{Additional arguments passed to \code{calc_dist}}
}
\value{
Character vector the same length as \code{x}
}
\description{
Standardise pain medication names
}
\examples{
drugs <- c("Paracetamol", "Paracetomol")
standardise(drugs)
}
